requirements.txt
google-generativeai
langchain-google-genai
langchain-community
langchain-core
langchain
langgraph
python-dotenv
streamlit
chromadb
pydantic
PyYAML
requests
rich
langchain-chroma
